third_party/WebKit/LayoutTests/fast/js/navigator-language.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description(
    "<pre>Check return value of navigator.language."
);

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setPOSIXLocale("en_US.iso88591");
}

var language = navigator.language.toLowerCase();
if (language.length == 2)
    shouldBe("language", "'en'");
else
    shouldBe("language", "'en-us'");
</script>
</body>
</html>
